TG: Add Sound lib

+++ NPEngine/Graphics/NPEngineGraphics.m
@@ -300,7 +300,7 @@ static NPEngineGraphics * NP_ENGINE_GRAPHICS = nil;
     return self;
 }
 
-- (unsigned)retainCount
+- (NSUInteger)retainCount
 {
     return UINT_MAX;  //denotes an object that cannot be released
 }
